Refactor the dispatch generator code to its own module
